+++ k/serial.h
@@ -1,9 +1,9 @@
 #ifndef KERNEL_EPITA_SERIAL_H
 # define KERNEL_EPITA_SERIAL_H
 
-#include <stddef.h>
+#include <k/types.h>
 
 void initSerial(int);
-int writeSerial(const void *str, size_t length);
+int writeSerial(const void *str, u32 length);
 
 #endif //KERNEL_EPITA_SERIAL_H
