+++ k/terminal.h
@@ -12,6 +12,6 @@ void setTerminalY(u16 y);
 void setTerminalColor(enum e_cons_codes fg, enum e_cons_codes bg);
 void writeTerminal(char c);
 void writeTerminalAt(char c, u8 color, u16 x, u16 y);
-int writeStringTerminal(const char *data, size_t size);
+int writeStringTerminal(const char *data, u32 size);
 
 #endif //KERNEL_EPITA_TERMINAL_H
